Change to floats, added loading code for all new syntax
It probably won't affect you at all, but worth noting that floats are
now stored as 2 32-bit values. I read the requirements again and it
seems like that's going to be required. My helper methods and the
loading code handle this right now, and I don't think anything other
than execution deals with that directly.